feat: created games repository. Used it in games bloc
