+++ twilight_imperium_tracker/lib/repository/GamesRepository.dart
@@ -1,0 +1,6 @@
+
+import 'package:twilight_imperium_tracker/feature/game/Game.dart';
+
+abstract class GamesRepository {
+  Stream<List<Game>> getGames();
+}
